+++ cli/cli/helpers/host_machine_directories/host_machine_directories.go
@@ -11,6 +11,10 @@ const (
 
 	sessionCacheFilename = "session-cache"
 
+	latestCLIReleaseVersionCacheFilename = "latest-cli-release-version-cache"
+
+	cacheFileForTests = "tests-content"
+
 	// ------------ Names of dirs inside Kurtosis directory --------------
 	engineDataDirname = "engine-data"
 )
@@ -33,6 +37,24 @@ func GetSessionCacheFilepath() (string, error) {
 		return "", stacktrace.Propagate(err, "An error occurred getting the session cache filepath from relative path '%v'", xdgRelFilepath)
 	}
 	return sessionCacheFilepath, nil
+}
+
+func GetLatestCLIReleaseVersionCacheFilepath() (string, error) {
+	xdgRelFilepath := getRelativeFilepathForXDG(latestCLIReleaseVersionCacheFilename)
+	latestCLIReleaseVersionCacheFilepath, err := xdg.CacheFile(xdgRelFilepath)
+	if err != nil {
+		return "", stacktrace.Propagate(err, "An error occurred getting the latest release version cache filepath from relative path '%v'", xdgRelFilepath)
+	}
+	return latestCLIReleaseVersionCacheFilepath, nil
+}
+
+func GetCacheFileForTest() (string, error) {
+	xdgRelFilepath := getRelativeFilepathForXDG(cacheFileForTests)
+	cacheFileForTestFilepath, err := xdg.CacheFile(xdgRelFilepath)
+	if err != nil {
+		return "", stacktrace.Propagate(err, "An error occurred getting the cache file for test filepath from relative path '%v'", xdgRelFilepath)
+	}
+	return cacheFileForTestFilepath, nil
 }